// src/modules/offering/income/enums/offering-income-creation-sub-type.enum.ts
export enum OfferingIncomeCreationSubType {
  SundayService = 'sunday_service',
  GeneralFasting = 'general_fasting',
  GeneralVigil = 'general_vigil',
  ZonalVigil = 'zonal_vigil',
  ZonalFasting = 'zonal_fasting',
  GeneralEvangelism = 'general_evangelism',
  ZonalEvangelism = 'zonal_evangelism',
  FamilyGroup = 'family_group',
  SundaySchool = 'sunday_school',
  YouthService = 'youth_service',
  TeenagerService = 'teenager_service',
  UnitedService = 'united_service',
  ZonalUnitedService = 'zonal_united_service',
  Special = 'special',
  Activities = 'activities',
  ChurchGround = 'church_ground',
}

export const OfferingIncomeCreationSubTypeNames: Record<
  OfferingIncomeCreationSubType,
  string
> = {
  [OfferingIncomeCreationSubType.SundayService]: 'Culto Dominical',
  [OfferingIncomeCreationSubType.FamilyGroup]: 'Grupo Familiar',
  [OfferingIncomeCreationSubType.GeneralFasting]: 'Ayuno General',
  [OfferingIncomeCreationSubType.GeneralVigil]: 'Vigilia General',
  [OfferingIncomeCreationSubType.ZonalFasting]: 'Ayuno Zonal',
  [OfferingIncomeCreationSubType.ZonalVigil]: 'Vigilia Zonal',
  [OfferingIncomeCreationSubType.GeneralEvangelism]: 'Evangelismo General',
  [OfferingIncomeCreationSubType.ZonalEvangelism]: 'Evangelismo Zonal',
  [OfferingIncomeCreationSubType.SundaySchool]: 'Escuela Dominical',
  [OfferingIncomeCreationSubType.YouthService]: 'Culto Jóvenes',
  [OfferingIncomeCreationSubType.TeenagerService]: 'Culto Adolescentes',
  [OfferingIncomeCreationSubType.UnitedService]: 'Culto Unido General',
  [OfferingIncomeCreationSubType.ZonalUnitedService]: 'Culto Unido Zonal',
  [OfferingIncomeCreationSubType.Activities]: 'Actividades',
  [OfferingIncomeCreationSubType.ChurchGround]: 'Terreno Iglesia',
  [OfferingIncomeCreationSubType.Special]: 'Especial',
};